requirements.txt
uvicorn==0.18.2
fastapi==0.79.0
python-multipart
whatimage==0.0.3
pyheif==0.7.0
Pillow==9.2.0